requirements.txt
argcomplete @ file:///tmp/build/80754af9/argcomplete_1618920853202/work
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569681205/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
Bottleneck==1.3.2
certifi==2021.10.8
cffi @ file:///tmp/build/80754af9/cffi_1636541929935/work
click==8.0.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz==0.11.0
dask==2021.10.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091426235/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
entrypoints==0.3
fonttools==4.25.0
fsspec @ file:///opt/conda/conda-bld/fsspec_1642510437511/work
imageio @ file:///tmp/build/80754af9/imageio_1617700267927/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1638529598726/work
ipykernel @ file:///tmp/build/80754af9/ipykernel_1633545415328/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1643818144432/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
jedi @ file:///tmp/build/80754af9/jedi_1644299024593/work
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter==1.0.0
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1636537035214/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver @ file:///opt/conda/conda-bld/kiwisolver_1638569886207/work
locket==0.2.1
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1634667019719/work
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune @ file:///tmp/build/80754af9/mistune_1594373098390/work
munkres==1.1.4
nbclient @ file:///tmp/build/80754af9/nbclient_1645431659072/work
nbconvert @ file:///opt/conda/conda-bld/nbconvert_1641309195684/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx @ file:///tmp/build/80754af9/networkx_1633639043937/work
notebook @ file:///tmp/build/80754af9/notebook_1644985983866/work
numexpr @ file:///tmp/build/80754af9/numexpr_1640686084435/work
numpy~=1.21.2
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.3.4
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partd @ file:///tmp/build/80754af9/partd_1618000087440/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1602788700959/work
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1636098896055/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2021.3
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658308664/work
PyYAML==6.0
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638434924971/work
qtconsole @ file:///opt/conda/conda-bld/qtconsole_1643819126524/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1643087291789/work
scikit-image==0.18.3
scipy @ file:///tmp/build/80754af9/scipy_1641555003412/work
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
shellingham @ file:///Users/ktietz/demo/mc3/conda-bld/shellingham_1629144685686/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
terminado @ file:///tmp/build/80754af9/terminado_1644304459949/work
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
tifffile==2020.10.1
toolz @ file:///tmp/build/80754af9/toolz_1636545406491/work
torch==1.10.2
torchvision==0.11.3
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typer @ file:///tmp/build/80754af9/typer_1637667998014/work
typing_extensions==4.1.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1645009354203/work
xmltodict @ file:///Users/ktietz/demo/mc3/conda-bld/xmltodict_1629301980723/work
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work

Pillow~=8.0.0
typer~=0.4.0
xmltodict~=0.12.0
matplotlib~=3.4.3
scikit-learn~=0.24.2
seaborn~=0.11.2